+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.6.0
+- Implement search command
+- Implement unused command
+
+## 0.5.0
+- Implement remove command
+- Run correct update command for Flutter project
+- Fix bug to ensure correct latest package version is pulled in
+- Update dcli package
+
+## 0.4.0+1
+- Update README
+
+## 0.4.0
+
+- Enable install particular package version
+- Enable adding dev dependencies
+
 ## 0.3.1
 
 - Fix missing pubspec.yaml bug caused when running `init` command in empty directory